Fix: move the wallet file only after closing it
Although our unofficial motto is "recreating Bitcoin's mistakes one at a time"
this particular case is a bit too much, even for us.

There's also the tiny detail of the pre-commit situation not working _at all_
in Windows, because that OS won't allow an open file to be moved (which is the
right thing).