Fix: don't wait for fresh chain while mining
Before this commit, we assumed that if the chain had become more than 5 minutes
old, we were on a stale chain, and started waiting for a fresh one. However, with
2-minute (desired, on average) block spacing a head of more than 5 minutes old
was in fact not unusual at all. Since all miners had this same condition built
in, this would mean the whole network would stop mining.

More fundamentally, the (beginner's) mistake here was to try to avoid waiste
for individual miners. In crypto:

* waiste is not something to avoid but to desire
* when forced to choose between the interests of individuals and that of the
  network we must always choose the network.

The bug has been fixed by:

* Taking the freshness-check out of the loop, and doing it on-start only.
* Increasing the bound for staleness to 40 minutes.