Networking: fix on self-hangup registration
We must only register the outgoing address to which we shouldn't connect. In
the previous situation, the outgoing address was never registered because the
hangup occurred on the incoming side. This led to many repeated connections
to self (which would all be short lived) and much cluttering of the logs.