Pre-fetch trusted blockchain when first starting the client
In true cryptocurrency fashion: if things don't work as promised/hoped, just
abandon all nice principles. In this case, rather than fix the mechanism for
"initial block download", just offer a download for the whole blockchain
somewhere and then just download automatically from there. (the principle being
abandoned, in part, is that of a peer to peer network)

Slightly more serious than the above: the mechanism for IBD has some room for
improvement, but I don't have the time for it right now, so the solution from
this post will probably work as a good workaround while we get all these new
joiners to the network.

Serious notes on security: if you just unpack blocks without verifying them,
you're running on an unverified chain. This is not something you should do in
general... in this case you're already trusting me, so little value is added in
re-verifying this particular chain, but if you wish you could do it from
skepticoin-repl by looping over all blocks and calling the various validate*
functions on each block in order.