Catch exceptions in "start_listening"
Before this fix, an error in start_listening (e.g. because another process
was already listening on 2412) would result in an easy-to-miss error on-screen,
followed by a non-running local peer. This would result e.g. in endless
"Waiting for fresh chain" messages. (You can wait till the cows come home but
without a running network thread you'll not get a fresh chain)